+++ CHANGELOG.md
@@ -7,9 +7,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 ## [Unreleased]
 
 ### Added
+- Added `Request.messages`, `Request.errors` and `Request.results`
 - Types for `SELECT.stream()`
+- `service.actions`
 ### Changed
 ### Deprecated
+- `service.operations`
+- Undocumented `service.entities()`, `service.events()`, `service.types()`, and `service.operations()`
 ### Removed
 ### Fixed
 - Types for `req.error()` et al.